test_requirements.txt
# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
mock
django-hvad
django-nose
coverage
django-coverage
django-mailer
fabric3
ipdb
mixer
factory_boy
flake8
beautifulsoup4
tox
